@@ -5,53 +5,51 @@
   "private": "true",
   "main": "server/index.js",
   "scripts": {
-    "dev": "webpack-dev-server",
-    "build": "NODE_ENV=production webpack",
-    "start": "node server/index.js"
+    "dev": "node server/index.js",
+    "build": "next build",
+    "start": "NODE_ENV=production node server/index.js",
+    "lint": "eslint .",
+    "precommit": "lint-staged",
+    "prettify": "prettier-eslint --write \"**/*.js*\" --list-different  --config .prettierrc"
   },
   "author": "MatthieuLemoine",
   "license": "MIT",
+  "lint-staged": {
+    "linters": {
+      "*.js": ["prettier-eslint --write --config .prettierrc", "eslint --fix", "git add"],
+      "*.json": ["prettier-eslint --write --config .prettierrc", "git add"]
+    }
+  },
   "devDependencies": {
     "babel-eslint": "^8.2.3",
-    "babel-loader": "^7.1.4",
     "babel-plugin-lodash": "^3.3.2",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "cache-loader": "^1.2.2",
-    "copy-webpack-plugin": "^4.5.1",
+    "babel-plugin-styled-components": "^1.5.1",
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.8.2",
-    "fast-async": "^6.3.7",
-    "file-loader": "^1.1.11",
+    "fast-async": "7.0.6",
+    "husky": "^0.14.3",
+    "lint-staged": "^7.1.2",
     "prettier": "^1.12.1",
-    "prettier-eslint-cli": "^4.7.1",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.8.3",
-    "webpack-cli": "^2.1.3",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.2"
+    "prettier-eslint-cli": "^4.7.1"
   },
   "dependencies": {
     "@sindresorhus/slugify": "^0.3.0",
     "algoliasearch": "^3.27.1",
     "chalk": "^2.4.1",
     "express": "^4.16.3",
+    "next": "^6.0.3",
     "nodent-runtime": "^3.2.1",
     "prop-types": "^15.6.1",
+    "qs": "^6.5.2",
     "ramda": "^0.25.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
     "recompose": "^0.27.0",
     "request": "^2.85.0",
     "request-promise-native": "^1.0.5",
     "styled-components": "^3.2.6"
-  },
-  "moduleRoots": [
-    "src"
-  ]
+  }
 }
